Work on fixing options binding
