+++ src/DMCopilot/DMCopilot.Backend/appsettings.Development.json
@@ -9,8 +9,7 @@
     "DataStore": {
         "Type": "CosmosDb",
         "CosmosDb": {
-            "EndpointUri": "https://dsr-dmcopilot-cdb.documents.azure.com:443/",
-            "DatabaseName": "dmcopilot"
+            "EndpointUri": "https://dsr-dmcopilot-cdb.documents.azure.com:443/"
         }
     },
     "SemanticKernel": {
