+++ src/DMCopilot/DMCopilot.Backend/appsettings.json
@@ -24,8 +24,7 @@
     "DataStore": {
         "Type": "CosmosDb",
         "CosmosDb": {
-            "EndpointUri": "https://dsr-dmcopilot-cdb.documents.azure.com:443/",
-            "DatabaseName": "dmcopilot"
+            "EndpointUri": "https://dsr-dmcopilot-cdb.documents.azure.com:443/"
         }
     },
     "SemanticKernel": {
